Initial SMP support.
This is a set of changes attempting to get basic SMP support going.
Tested with qemu and secondary CPUs are brought up using PSCI.
Secondary CPUs are coming up as expected and are receiving and handling
interrupts.

This is buggy at present. For example with qemu, there are
arbitrary faults hitting the secondaries, usually when they are
running user-space code.

Hopefully this work will act as a stepping stone for incremental
additions through to proper SMP support.

Conflicts:
	sys/arm64/arm64/locore.S
	sys/conf/options.arm64